Refactored handling of values/valuesets entirely
ValueSets and "values" are now not handled differently anymore. This
simplifies code, improves readability, adds some cool features and,
generally, would have been a good idea from the get-go. Handling the
only case that justified having the two separate (namely, having an
order to the list of values) is now taken care of by the OrderedSet
class, which supports both keeping the original order, but also having a
sorted version of elements internally.