Fix numerical error in unit conversion
With values in the range of 0 < x < 1, there was a numerical error when
converting e.g. "1u". It was previously converted to a rounded floating
point mantissa value of 999.99999999, leading to re-conversion as
"1000n". Fix this by calculating with exact numbers all the way during
the conversion and add a specific testcase for that issue.